EndgunTableEntry, EndgunTable, EndgunSetup
The EndgunTableENtry and EndgunTable classes enable the description of two behaviors of an endgun: its throw vs pressure, and flow vs pressure functions. Used in EndgunSetup, which describes the particular endgun being used. It's used by value in IrrSystemConfiguration (as opposed to being a descendant of DeviceElementConfig) because PAIL considers endguns a special kind of section, and endgun setup data is not always available.