Adding  ObsCollection reference to Load
Supporting the use case where a set of loads (e.g., a USDA report coming ffrom a cotton gin that describes multiple bales; a set of railroad cars loaded with grain coming to an elevator) has a collection of observations per load (e.g., staple length, micronaire, color, etc. for cotton; moisture, protein content, etc. for grain). It's natural for all the observations in that shipment to fall within a single Observations document (already supported), but it's reasonable for each Load's observations to be held within a specific ObsCollection, so the observations for each Load can be told apart. The reference to an ObsCollectionId added in this pull request enables this.